Move to SDI>=0.4 from sborl. Consolidate into one file.
- Migrate from 'sborl' to new SDI release
- Move library into one file
- Pack schema as a dict into the relation library
- Fix the tests
