--- lib/charms/traefik_k8s/v0/ingress_per_unit/schema.yaml
@@ -1,40 +1,0 @@
-v1:
-  requires:
-    unit:
-      type: object
-      properties:
-        model:
-          type: string
-        name:
-          type: string
-        ip:
-          type: string
-        port:
-          type: integer
-        prefix:
-          type: string
-        rewrite:
-          type: string
-      required:
-        - model
-        - name
-        - ip
-        - port
-        - prefix
-        - rewrite
-  provides:
-    app:
-      type: object
-      properties:
-        ingress:
-          type: object
-          patternProperties:
-            "":  # unit name
-              type: object
-              properties:
-                url:
-                  type: string
-              required:
-                - url
-      required:
-        - ingress
