+++ requirements.txt
@@ -1,4 +1,4 @@
 ops >= 1.3.0
 lightkube >= 0.8.1
 lightkube-models >= 1.22.0.4
-sborl<1.0.0
+serialized-data-interface>=0.4.0
