+++ tests/unit/test_lib_helpers.py
@@ -2,14 +2,13 @@
 # See LICENSE file for licensing details.
 """Helpers for unit testing charms which use this library."""
 
-from ops.model import Relation
-from sborl.testing import MockRemoteRelationMixin
-
-from .ingress_per_unit import (
+from charms.traefik_k8s.v0.ingress_per_unit import (
     IngressPerUnitProvider,
     IngressPerUnitRequirer,
     IngressRequest,
 )
+from ops.model import Relation
+from serialized_data_interface.testing import MockRemoteRelationMixin
 
 
 class MockIPUProvider(MockRemoteRelationMixin, IngressPerUnitProvider):
