+++ tests/unit/test_lib_requires.py
@@ -6,10 +6,10 @@ from textwrap import dedent
 from unittest.mock import Mock
 
 from charms.traefik_k8s.v0.ingress_per_unit import IngressPerUnitRequirer
-from charms.traefik_k8s.v0.ingress_per_unit.testing import MockIPUProvider
 from ops.charm import CharmBase
 from ops.model import Binding
 from ops.testing import Harness
+from test_lib_helpers import MockIPUProvider
 
 
 class MockRequirerCharm(CharmBase):
